@@ -5,6 +5,8 @@
 	"license": "MIT",
 	"repository": "vweevers/vfile-reporter-shiny",
 	"author": "Vincent Weevers",
+	"type": "module",
+	"exports": "./index.js",
 	"scripts": {
 		"test": "standard && depcheck && ava && hallmark"
 	},
@@ -40,6 +42,6 @@
 		"serial": true
 	},
 	"engines": {
-		"node": ">=10"
+		"node": ">=12.20"
 	}
 }
